refactor: move impure package to compat
