+++ writer.go
@@ -1,4 +1,4 @@
-package impure
+package lipgloss
 
 import (
 	"fmt"
@@ -7,6 +7,10 @@ import (
 
 	"github.com/charmbracelet/colorprofile"
 )
+
+// Writer is the default writer that prints to stdout, automatically
+// downsampling colors when necessary.
+var Writer = colorprofile.NewWriter(os.Stdout, os.Environ())
 
 // Println to stdout, automatically downsampling colors when necessary, ending
 // with a trailing newline.
